LF-12838: vvcam: Fix the os08a20 sensor capture failed issue
When use the os08a20 do capture on imx8mp, the camera can't work.

The issue is caused by:
commit <657cc0dc3996> (media: ox05b1s: Add support for OS08A20 raw sensor)

The root cause is that the os08a20 driver in the vvcam is not found
correctly. Both of the compatible name "ovti,os08a20" and
"ovti,os08a20_mipi" in this two commits are used for other SOC, and if use
the same compatible name for imx8mp, that leaded the VSI isp camera create
pipeline failed. So need to rename the compatible name of os08a20 in
os08a20_mipi_v3.c file to fix this issue.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>